+++ ArmStuff/ArmMain/Potentiometer.cpp
@@ -1,8 +1,8 @@
 #include "Arduino.h"
 #include "Potentiometer.h"
 Potentiometer::Potentiometer(int _pin){
-	pin = _pin;
+  pin = _pin;
 }
-Potentiometer::getValue(){
-	return analogRead(pin);
+int Potentiometer::getValue(){
+  return analogRead(pin);
 }
