+++ ArmStuff/ArmMain/Potentiometer.h
@@ -2,10 +2,10 @@
 #define Potentiometer_h
 #include "Arduino.h"
 class Potentiometer{
-	public:
-		Potentiometer(int _pin);
-		int getValue();
-	private:
-		int pin;
+  public:
+    Potentiometer(int _pin);
+    int getValue();
+  private:
+    int pin;
 };
 #endif
