--- ArmStuff/ArmMain/libraries/Encoder.cpp
@@ -1,11 +1,0 @@
-#include "Arduino.h"
-#include "Encoder.h"
-
-Encoder::Encoder(int A, int B){
-	pinA = A;
-	pinB = B;
-	countValue = 0;
-}
-Encoder::getEncoderValues(){
-	return countValue;
-}
